Update Client side to call server port 8082
